Refactor and added BMP180 Temp
Refactored I2c to work from i2c.c file. This allows the task to focus on processing the sensor data. BMP180 calibration and Uncomp. Temp. is read and then the actual temp is calculated. There is a BMP180 test code that executes the actual T and P functions using data provided in the datasheet. This allows confirmation that the algorithms are working as expected.